+++ package.json
@@ -5,13 +5,19 @@
   "description": "Личный проект «Большое путешествие» от HTML Academy",
   "scripts": {
     "eslint": "eslint src/**/*.js",
+    "start": "webpack-dev-server --open",
+    "build": "webpack",
     "test": "npm run eslint"
   },
   "devDependencies": {
-    "eslint": "6.1.0",
-    "eslint-config-htmlacademy": "0.5.2"
+    "eslint": "5.16.0",
+    "eslint-config-htmlacademy": "0.5.2",
+    "webpack": "4.42.1",
+    "webpack-cli": "3.3.11",
+    "webpack-dev-server": "3.10.3"
   },
   "engines": {
     "node": "10"
-  }
+  },
+  "dependencies": {}
 }
