+++ src/components/days.js
@@ -1,0 +1,8 @@
+const createTripDays = () => {
+  return (
+    `<ul class="trip-days">       
+    </ul>`
+  );
+};
+
+export {createTripDays};
