+++ src/components/trip-info-cost.js
@@ -1,0 +1,9 @@
+const createTripInfoCost = () => {
+  return (
+    `<p class="trip-info__cost">
+      Total: &euro;&nbsp;<span class="trip-info__cost-value">0</span>
+    </p>`
+  );
+};
+
+export {createTripInfoCost};
