+++ src/components/trip-info.js
@@ -1,0 +1,8 @@
+const createTripInfo = () => {
+  return (
+    `<section class="trip-main__trip-info  trip-info">     
+    </section>`
+  );
+};
+
+export {createTripInfo};
